Re-enable runtime badging of browser windows.
This CL fixes an issue with the profile icon not showing up if the user
has pinned the default chrome shortcut to the taskbar.

Using the runtime badging on top of the taskbar icon's badge causes the badge to look slightly darker around the edges at runtime.

BUG=313800

Review URL: https://codereview.chromium.org/54053008