remove CHECK in RemoteRootWindowHostWin::Disconnected
This was triggered by the change to restart in desktop mode when running
with the software compositor. But in general, it's possible for a channel
error to happen before SetTargetSurface which sets the host_. Since the
CHECK isn't accomplishing too much, just remove it.

R=cpu@chromium.org, sky@chromium.org
BUG=315057

Review URL: https://codereview.chromium.org/62853002