Add flow to enable Google Drive offline mode automatically on first run.
How this works is that we create a WebContents in the background to a Drive
endpoint that will take care of all the offline initialization for the app.
We succeed if a background page is opened; otherwise, we fail if the page does
not load or after a timeout.

BUG=307302
TEST=manual

Review URL: https://codereview.chromium.org/55423004