Add VDA decode time test and refactor.
This measures the median of the decode time of all the frames.
VDA::Decode is called 30 times per second to simulate WebRTC.

VideoDecodeAcceleratorTest class is renamed to
VideoDecodeAcceleratorParamTest, which shows it is a
value-parameterized test. The common code is moved to a new
test fixture VideoDecodeAcceleratorTest, which is inherited
by VideoDecodeAcceleratorParamTest. The future new tests should
be based on VideoDecodeAcceleratorTest if they do not need to
test different combinations of the parameters.

BUG=292703
TEST=Run VDA tests.

Review URL: https://codereview.chromium.org/57663002