Refactor SyncBackendHost
Split SyncBackendHost into SyncBackendHost, SyncBackendHostImpl and
SyncBackendHostCore.

The SyncBackendHost class is now a pure virtual interface.  It is the
contains the interface exposed to the ProfileSyncService.

The SyncBackendHostImpl class implements the functionality of the
SyncBackendHost.  It contains most of the code that used to reside in
SyncBackendHost.

The SyncBackendHostCore is the old SyncBackendHost::Core under a new
name.  It's been moved out of the SyncBackendHost's .cc file and into
its own, publicly declared class.  This will hopefully enable better
testing in the future.

This is a pure refactoring change.  It should have no impact on program
behavior.

BUG=312999

Review URL: https://codereview.chromium.org/61183003