Allow printing component to be removed completely
Using enable_printing=0 removes all of the source files of the printing
component, but has no effect on its dependencies. It is wrong to empty out the
target inside the target. If printing is disabled, then the dependency on
printing.gyp is not needed and can be removed.

This target was causing libgcrypt to be linked even when printing was disabled.
That problem goes away with this change.

Review URL: https://codereview.chromium.org/62793002