Password manager: Fix focus issue with show/hide button.
Instead of updating the model when revealing a password, which causes
the UI to flash as well as causing the button to lose focus, simply
update the DOM element directly. This is probably the right thing to
do anyways, otherwise a password that was revealed once would remain
accessible in the DOM as long as the tab remains open, which is not
what we want.

BUG=315491,309553
R=bauerb@chromium.org

Review URL: https://codereview.chromium.org/59233010