Added a BCKS to own a Profile's SchemaRegistry.
The SchemaRegistry contains information about the policy types that should be
loaded, and is a dependency of the policy providers. The new BCKS allows
creating the Profile's SchemaRegistry before the Profile's policy providers are
created and start providing Values to the PrefService.

Also added a global SchemaRegistry at the BrowserPolicyConnector that merges
the registries of all the Profiles. This enables shared policy providers (e.g.
the platform provider) to load policy for all the components of all the
Profiles.

R=bartfab@chromium.org, erg@chromium.org, phajdan.jr@chromium.org
TBR=joi@chromium.org
BUG=270667

Review URL: https://codereview.chromium.org/50493014