Added a purge by predicate method to the ResourceCache.
This enables removing stale entries by passing a callback that tests if
a key is still valid or not, instead of having to put all the currently good
keys in a set. The cache user can now wrap a custom function that works with
whatever structured contains the current keys to keep.

BUG=270667
R=bartfab@chromium.org

Review URL: https://codereview.chromium.org/52553005