Convert all synthetic gesture coordinates from CSS pixels to DIPs.
There is currently some inconsistency about the units of pixels being passed to
synthetic gestures, some are CSS pixels and some are density independent pixels
(DIPs). The internal classes expect DIPs but the scroll distance is being
passed in CSS pixels, causing us to scroll too far on pages with page scale
factors > 1.

This patch clarifies the interface between the JS functions for gestures and the
internal classes. The JS functions expect coordinates to be in CSS pixels which
are then converted to DIPs before being passed to the gesture objects.

BUG=313669

Review URL: https://codereview.chromium.org/56093002