Disable the panel tests.
During the last attempt to switch to openbox, the panel tests started
timing out. Given that we hope to switch the main builders to linux_aura
this weekend, and that all of the panel tests are disabled there, we
shouldn't mind losing coverage on a port we aren't going to ship for a
couple of days.

BUG=315326
R=kbr@chromium.org
TBR=kbr@chromium.org

Review URL: https://codereview.chromium.org/64633004