Add cross-platform API to get the form factor of the device
This API returns whether the device is a phone or a tablet for Android and iOS
(for other platforms, it always returns the device as being desktop). The new
API is placed in //ui rather than //chrome in order to make it accessible to
components.

BUG=312903

Review URL: https://codereview.chromium.org/50603005