Refactor ProfileSyncServiceStartup tests
This CL removes the ProfileSyncServiceStartupTest's dependency on the
TestProfileSyncService.  It does this by using a real ProfileSyncService
instead, with the help of some injected mock components.  In particular,
the tests now make use of a SyncBackendHostMock, a class which is
introduced in this CL.

This CL also refactors FakeOAuth2TokenService into its own file.  This
allows us to make it available to
profile_sync_service_startup_unittest.cc without having to include
test_profile_sync_service.h.

BUG=31299

Review URL: https://codereview.chromium.org/65383003