Linux sandbox: make BPF policies an implementation of an interface.
This is the first pass. We renamed SetSandboxPolicy() to
SetSandboxPolicyDeprecated().

The new SetSandboxPolicy() takes a pointer to an instance of a class that
implements a simple interface.

This will improve flexibility by removing the need to pass static pointers
for policies. In particular, this will allow us to greatly improve
readability of the content/ policies and to use C++ inheritance to implement
our policy hierarchy.

BUG=316486
R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/62953011