Eliminate Shell::RootWindowList in favor of aura::Window::Windows.
This is necessary as part of my plan to rework RootWindow, specifically making it not be a subclass of Window. (It will eventually be renamed WindowEventDispatcher). Code that expected to get the RootWindow interface from this type can still call GetDispatcher() on members of the returned vector of Windows.

R=oshima@chromium.org
http://crbug.com/308843

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=233787

Review URL: https://codereview.chromium.org/64933002