Added VideoCaptureDeviceAvFoundation class, implementing the PlatformVideoCapturingMac protocol.
All the protocol methods are NOTIMPLEMENTED(), the static +getDeviceNames()
is implemented though and used to enumerate the devices in the system in 
VideoCaptureDeviceMac if IsAVFoundationSupported()==YES. 

The rest of the VideoCaptureDeviceMac operations use QTKit still.

This CL is part of a larger exercise to add support for Video Capture in 
Mac > 10.6 using AVFoundation. 

BUG=288562

Review URL: https://codereview.chromium.org/48983002