Get rid of user-level styles.
-The Apps codepath for this is just using the wrong thing.
It should be using author-level styles like extensions do.
-The user-stylesheet feature requires the user to put a CSS
stylesheet in the right location in their user-data-dir.
Extensions are a much better way of doing this.

This is in preparation for simplifying the Blink style
resolution code by removing the concept of user styles.

Review URL: https://codereview.chromium.org/64843004