Forcing the EABI hard-float attribute off.
GCC won't link chacha_vec_arm.s with the following error:
  "uses VFP register arguments, output does not".
This appears to be because the .s file specifies hard-float ABI and the rest
of the system is built with soft-float. Since the function in this .s file
doesn't take or return any floating point arguments, this change forces the
EABI flag off in the hope that it'll all work without crashing.

(I don't have a build setup, so this is development-by-buildbot.)