Don't Generate ui::TouchEvent from xevent if it does not belong to a RootWindow
When generating an ui::TouchEvent from an xevent, it needs to consult the
singleton DeviceDataManager & TouchFactory, and may also change their internal
states, e.g. a touch release event can clear its touch id in the record.
Thus converting a same ui::TouchEvent into an xevent multiple time can
cause problem. This can currently happen when we have multiple displays and
each display is associated with a RootWindow. An X touch event is dispatched
to each RootWindow, converted to ui::TouchEvent, and then it is decided
whether the TouchEvent belongs to the particular RootWindow.

This CL changes that when the X touch event is dispatched to RootWindow,
bails out early if the event does not belong to the RootWindow, before
converting it to ui::TouchEvent. So it makes sure that the X event is
only converted to ui::TouchEvent once, for the right RootWindow.

BUG=315840
TEST=Connect Pixel to external touchscreen. Turn on touch HUD projection.
     Do multi-finger gestures on external touchscreen and make sure all
     touch points are released correctly.

Review URL: https://codereview.chromium.org/63043004