This CL introduces a lightweight bindings system for V8 called gin
Unlike the extensions V8 bindings, gin is based on ObjectTemplates rather than
on evaluating script.  Unlike the Blink V8 bindings, gin isn't tightly coupled to
Blink.  In fact, gin's only link-time dependency is V8.  We plan to use gin to build
the V8 bindings for Mojo (see https://codereview.chromium.org/59153005/ for
an example of how they will be used).

In the long term, gin could serve as a basis for both the Blink and the extension
system bindings, but we don't have any immediate plans to pursue that use of
this code.

This code is largely inspired by a lightweight bindings system designed by
Aaron Boodman.

Review URL: https://codereview.chromium.org/67763002