chromium-status: tighten up the Who/When columns
Often times the Message field is very long (especially when multiple
failures are logged in there). This forces the other two columns to
get wrapped heavily. For the date, this makes the field unreadable:
	Sat
	Nov
	02,
	17:09
	UTC

Shrink the size slightly for both these fields (as we just want to
scan them vertically, but otherwise focus on the Message), disable
white space breaking on the date, and move the redundant timezone
to the column header.

BUG=chromium:314501
TEST=ran an appengine instance and checked the output
TEST=`./tests/main_test.py` passes

Review URL: https://codereview.chromium.org/57643007