chromium-status: make the limits field discoverable
People don't really know about the limit portion of the URL to load up
more results.  Add a link to the existing main line, and make sure the
number reflects the reality of how many entries we are showing.  It's a
cheap way for users to notice things.

BUG=chromium:314501
TEST=ran an appengine instance and checked the limit behavior/message
TEST=`./tests/main_test.py` passes

Review URL: https://codereview.chromium.org/58353002