Move Archive class to scripts/slave/chromium
The Archive class is only used in a few places, and should live closer to those
scripts that use it. This CL helps clean up site_config for a larger refactoring
in the near future. This CL also has an internal component that must be landed
at the same time.

R=maruel@chromium.org

Review URL: https://codereview.chromium.org/60153003