Adding support for forced URLs to TopSites.
This CL makes it possible to add MostVisitedURLs to TopSites that have a non-null |last_forced_time| indicating that these URLs are "forced" and should stay in top-sites even as other non-forced URLs come in. These forced URLs are evicted on a Least Recently Used basis based on |last_forced_time|. Internally, the list of Forced URLs is kept at the beginning of the TopSitesCache list of MostVisitedURLs. These forced URLs are not returned by GetMostVisitedURLs.

Forced URLs make it possible for Chrome to capture thumbnails of sites that may have been suggested through other processes than the TopSitesBackend. They are not used yet but a soon-to-follow CL will link them to thumbnails requested via the chrome-search://thumb2/...

TBR'ing owners for minor API change.

BUG=
TBR=bulach@chromium.org,sky@chromium.org,darin@chromium.org,benwells@chromium.org,estade@chromium.org,

Review URL: https://codereview.chromium.org/53283004