Rename fileBrowserPrivate.logoutUser to logoutUserForReauthentication.
The method is used for navigating the user to the GAIA login
screen (which we can only during the account log-in flow for now on
Chrome OS), when for some reason Files.app detected token invalidation.

This CL renames the private API to clarify its purpose.
In later CLs, I'd like to ensure that the API does navigate
the user to the the desired login flow, for fixing 307642.

BUG=307642
TBR=isherman@chromium.org

Review URL: https://codereview.chromium.org/68343003