Don't use UpdateRegistryFromResourceS as it's not in 2013
ATL version that ships with 2013 doesn't have a suffix. 
UpdateRegistryFromResource used to select either an S or
a D suffix based on the value of _ATL_STATIC_REGISTRY,
however in ATL 12 (vs2013), UpdateRegistryFromResource
now just means UpdateRegistryFromResourceS, and
UpdateRegistryFromResourceD is separate.

We want to keep building with 2010 and 2013, so #if based
on _ATL_VER.

R=ananta@chromium.org, robertshield@chromium.org
BUG=309197

Review URL: https://codereview.chromium.org/66583008