Make ui::SelectionRequestor robust to recursive message loops.
This was causing a hang in BookmarksTest.CommandOpensBookmarksTab. In
particular, multiple instances of ui::SelectionRequestor would end up on
the stack, and the ones further up the stack would discard the results
intended for the instances further down the stack.

This keeps an explicit list of pending requests and fulfills them in
FIFO order.

BUG=315831

Review URL: https://codereview.chromium.org/66843005