Policy providers all get a SchemaRegistry to work with.
The SchemaRegistry contains the components that the policy provider should load
policy for, as well as the expected policies and their types for each component.

Also killed the PolicyDomainDescriptor, which was a similar way of doing this
but had the problem of overlapping descriptors on shared policy providers.

BUG=270667
R=bartfab@chromium.org, jochen@chromium.org, kalman@chromium.org

Review URL: https://codereview.chromium.org/56623005