Removed the PolicyDefinitionList.
This list was generated at compile time and included policy information
extracted from the policy templates file that was useful at runtime, such as
the policy names, types, deprecation status, etc.

Most of this information is now available via the Chrome policy schema, so
this list was largely redundant.

The chrome-specific data in this list is now available in a list of
PolicyDetails, also generated at compile time and kept in sync with the Chrome
schema data.

Refactored all the remaining users of PolicyDefinitionList to use the schema
and/or PolicyDetails now. Also introduced a test helper to mock out
PolicyDetails in tests.

Cleanup: moved policy_test_utils.h to policy/test/

BUG=270667
R=bartfab@chromium.org, dubroy@chromium.org

Review URL: https://codereview.chromium.org/58313002