x11: Create mouse-wheel events properly when XI2.2 isn't being used.
X11 sends button press/release events for mouse-wheel events. It is necessary to look
at the button to decide whether the event is a normal mouse-press/release event, or
whether it is a wheel event. We do this correctly for XI2.2 events. This fixes the code
when X11 sends non-XI2.2 events.

BUG=317286
R=erg@chromium.org

Review URL: https://codereview.chromium.org/68513012