cc: Combine analysis and raster
This changes the analysis of PicturePiles from optimistic (hoping to
save raster work by checking if a rastered bitmap would be a solid
color) to pessimistic (assuming most bitmaps are not solid colors and
just checking afterwards).  It does this by rastering to both an
AnalysisCanvas and a normal SkCanvas via SkNWayCanvas and then
checking the results afterwards.

R=reveman@chromium.org, senorblanco@chromium.org
BUG=310796

Review URL: https://codereview.chromium.org/63443003