Pepper: Tighten GetLiveModule in PluginRegistry.
On PepperPluginInstanceImpl::Delete(), we send a synchronous DidDestroy message
to the plugin.  The plugin can exit() in response to this message.

While the renderer is waiting for a response (and holding a reference to
PepperPluginInstanceImpl on the stack), another plugin for the page could
create an instance of a plugin using the same path as the deleted instance.

This can cause a PluginModule to be reused on the host for a plugin which is no
longer live. In this case, sending DidCreate will fail.

Alternatively, the lifetime of PluginModule could be cleaned up, but there are
many uses of PepperPluginInstanceImpl::module() throughout the codebase. I'm
happy to look at seeing if that can be cleaned up in another change.

BUG=
R=dmichael@chromium.org, teravest@google.com

Review URL: https://codereview.chromium.org/69363002