Don't lock the cursor when dragging or resizing a window with touch
The user is already able to simultaneously move the 
mouse cursor while touching the screen (e.g., 
while touch-scrolling web contents,reordering 
tabs with touch, etc.), so touch-dragging or resizing 
a window should be no exception. Do not lock the 
cursor when initiating a touch drag/resize of a 
browser window so that a physical mousemove will 
show the cursor.

BUG=316654
TEST=WorkspaceWindowResizerTest.MouseMoveWithTouchDrag

Review URL: https://codereview.chromium.org/67403003