cc: don't lose resources sent to parent when ResourceProvider::Child is destroyed.
Since we recreate the DelegatedRendererLayer in some cases (e.g. child changed,
size changed), we'll delete the Child, but we don't want to lose resources that
may be in the parent. Instead we can delay the Child destruction until all
resources have been returned (or we destroy the ResourceProvider).

BUG=315847

Review URL: https://codereview.chromium.org/66803006