Initialize the views::WebView's preferred size in content_shell
When running layout tests we want the web contents area to be set to the
shell's initial size (almost always 800x600). This sets the preferred size
on the views::WebView when initializing the web contents so the views layout
manager sizes us correctly. The code was previously trying to size the root window,
which results in the web contents area being slightly smaller as the layout
manager was reserving space for UI elements.

(Re-uploaded from 67193003, initial patch b/ jamesr).

TBR=jochen@chromium.org
BUG=309760

Review URL: https://codereview.chromium.org/69903003