Mojo: Implement local passing of MessagePipes.
I.e., allow handles to MessagePipes to be transferred across
MessagePipes whose ends are both local (i.e., in-process).

There's one basic test of this functionality in this change, but much
more is needed.

Also, in theory, a handle to a remote (proxied) MessagePipe can be
transferred across a local MessagePipe, but this is not tested.

R=davemoore@chromium.org, davemoore

Review URL: https://codereview.chromium.org/66963005