Android: keep a "heartbeat" for perf tests.
The script's output is "buffered" so it can be retried on case of failures
without messing up the higher level log parsing.
This however triggers the buildbot watchdog, which kills the tests since
it doesn't produce any output.
Print a hearbeat message to indicate the --single-step step script is
alive and kicking.

BUG=318369
NOTRY=True

Review URL: https://codereview.chromium.org/68503009