linux_aura: Minimal implementation of window icons.
This implements the X11 side of window icons, setting the _NET_WM_ICON
property on the x11 window. This also quickly hacks a path to setting the chrome icon.

This icon is obviously upscaled in some contexts. Actually fixing this
requires redoing a bunch of plumbing inside of views so that vectors of
ImageSkias are passed around so that we can offer the X server all
representations that we know about instead of two, hardcoded, Windows
sizes.

BUG=317888, 318466
R=ben@chromium.org

Review URL: https://codereview.chromium.org/70183005