Added permission for chrome.app.window alwaysOnTop property
Apps that wish to enable the alwaysOnTop property for app windows must
now declare the "alwaysOnTopWindows" permission in their manifest.
This permission does not appear in install prompts.

BUG=314260, 171597
TEST=browser_tests (PlatformAppBrowserTest.WindowsApiAlwaysOnTop*).
Manual test: Create an app which does not specify this permission and
attempts to create a new window with the option enabled or calls
setAlwaysOnTop(true) on an existing window. The window should not be
always on top. The option should work for apps which do declare the
permission.

Review URL: https://codereview.chromium.org/68423003