Make the "Consent To Send Stats" file follow the kMetricsReportingEnabled pref.
The options UI updates the state of this file manually (besides setting the
pref), but when the policy overrides the pref then the file should be updated
to reflect the pref's value.

On Windows a registry key is kept in sync with the pref instead of the consents
file.

After this fix it'll be possible to simply set the pref from the options UI,
and to remove the ChromeOS-specific code (that reads from device settings).

BUG=318761

Review URL: https://codereview.chromium.org/70233005