Allow PPB_ImageData_Impl to be created in a way that is compatible with unittests
On some platforms, PPB_ImageData_Impl needs to send a message to the browser to create
its transport dib. For unittests, this doesn't work and there is no sandbox so the
transport dib can be created directly. This adds a construction flag to PPB_ImageData_Impl
which allows it to be created in a way that works in unittests.

BUG=
R=bbudge@chromium.org

Review URL: https://codereview.chromium.org/61813003