[android] Remove stale chromium_org link before sync.
If a symlink to the "regular" Chromium source is present under
android-src/external/chromium_org and the upstream Android tree
contains a clone of the Chromium source code in that location
repo will attempt to sync the "regular" Chromium source code to
revisions which are only present in the Android checkout.
The fix is to remove the symlink before running repo sync.

BUG=313516

Review URL: https://codereview.chromium.org/68763010