Add SetOffset to PPB_Graphics2D_Dev
Add a SetOffset function to PPB_Graphics2D_Dev which allows the graphics context to be offset into the plugin element. This allows having a graphics context which is smaller than the plugin element and painting into only a portion of the plugin elements. This makes it possible to do viewporting for very large plugins.

R=piman@chromium.org, tsepez@chromium.org, yzshen@chromium.org

Review URL: https://codereview.chromium.org/25661002