Allow Graphics2D stretch the plugin contents when resizing the plugin.
When a plugin element is resized, the new size is asynchronously sent from the
renderer to the plugin process. There is a delay between the plugin
being resized and a new plugin frame being drawn by the plugin. The current
behavior is to leave the current frame on the screen until a new frame is
received. This results in a flickering effect while the plugin draws the new
frame.

Many plugins stretch their contents to fill the given plugin element. With this
change, plugins have the option to allow the last frame painted to the renderer
to be stretched in proportion to
how much a plugin element has changed in size since it received its last
Flush(). This eliminates the flickering effect in these cases.

The setting is provided under a PPB_Graphics2D_Dev SetResizeMode function.

BUG=
R=piman@chromium.org, tsepez@chromium.org

Review URL: https://codereview.chromium.org/57613002