Make ResourceHandler::OnResponseCompleted take a defer output parameter.
Every other method of ResourceHandler returns false to abort and sets *defer to
cancel. Between the StreamResourceHandler fix here and
https://src.chromium.org/viewvc/chrome?view=rev&revision=233661, there were
twice as many ResourceHandlers that mistakenly deferred shutdown as
intentionally. (RedirectToFileResourceHandler and CrossSiteResourceHandler do
so intentionally.) Make this consistent so it's not as confusing.

BUG=none

Review URL: https://codereview.chromium.org/74373002