Refactor some ProfileSyncServiceTests
Create a new test framework class named ProfileSyncServiceSimpleTest and
use it to replace several ProfileSyncServiceTests.  This newer framework
takes advantage of MockSyncBackendHost and allows us to use a real
ProfileSyncService in these tests.

The resulting tests are better at testing the ProfileSyncService, but
they no longer test any part of the SyncBackendHost code.   This is a
desirable trade off because these are unit tests.

BUG=312994

Review URL: https://codereview.chromium.org/69583002