Set focus to the "OK" button when editing a supervised user profile
Since a supervised user can't edit their name, trying to set the focus there
resulted in it ending up on the first element in the tab order. On Mac and
Linux, that was the "Cancel" button, rather than "OK", because the button order
is reversed to fit those platforms' conventions. Set focus more deliberately.

BUG=315532
TEST=new browser_test

Review URL: https://codereview.chromium.org/75363002