Implement Mojo message codec in JavaScript
This CL contains an encoder and decoder for the Mojo message IPC format.  I've
tested the encoder and decoder against each other, but I haven't yet tested them
against the C++ implementation.

I haven't tried to tune the encoder or decoder for performance. IMHO, we should
work on that after getting correctness.

The C++ code for JSTest.codec is a bit verbose because the C++ code need to
load all the dependencies for codec_unittests.js. A better approach is to teach
the system how to load dependencies from the file system, but I've left that for
a future CL.

BUG=317398
R=aa@chromium.org

Review URL: https://codereview.chromium.org/69843003