Use scoped_ptr to clarify the ownership of the returned object
ui::CreateInputMethod transfers the ownership of the returned
object to the caller.  It should use scoped_ptr<InputMethod>
instead of asserting the ownership in the header comment.

Note:
This CL does not touch GetSharedInputMethod and
DestroySharedInputMethod because they will be removed as
filed in Issue 319122.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/74613002