Support headless content_shell for linux/aura.
When running blink layout tests, we shouldn't create native windows. We should
rather just create the virtual content window for render of content. Headless
mode of content shell support is already present on other ports and needs to be
added for aura. It is important in terms of performance (when running layout tests).

BUG=316712

R=jochen

TEST=None; no behavior change;

Review URL: https://codereview.chromium.org/71243002