Fix reporting of Extensions.KnownDisabledReDisabled to also report 0 when there are no known_disabled extensions.
As it is now we don't report neither Extensions.KnownDisabledReDisabled:0, nor Extensions.KnownDisabledInitialized, when there are no known_disabled and no currently disabled_extensions.

This CL makes it so that we always report either and will result in an increase of the number of 0 reports for Extensions.KnownDisabledReDisabled.

Review URL: https://codereview.chromium.org/76533002