Add a gin/public directory
When we start using gin from other projects but mojo, we want to be
extra careful what we include. Therefore, we introduce a public/
directories with headers that are ok to be included from gin-embedders.

As a start, we move wrapper_info.h into public/.

BUG=317398
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/69293015