Ensure that the first browser window which is created in Windows 8 ASH is maximized.
This is done via the Ash WindowPositioner class. The change is in the WindowPositioner::GetBoundsAndShowStateForNewWindow
function which for windows returns SHOW_STATE_MAXIMIZED if this is the first window being opened.

Based on a discussion with oshima, I enabled the WindowSizerAshTest ash unit tests for Windows. There were a number
of places in the WindowSizer class where we were allowing browsers to be in ASH only for Chrome OS. Checking for
the host desktop type should be enough. This was needed to get the window to be maximized for ASH.

Fixes bug http://crbug.com/231603

BUG=231603
R=cpu@chromium.org, oshima@chromium.org, sky@chromium.org, cpu, oshima, sky

Review URL: https://codereview.chromium.org/62333015