Cancel all in-progress downloads when the last browser window is closed.
When quitting using wrench quit or when the last browser window is
closed on Windows and Linux, a dialog warns that all downloads will be
cancelled. Previously, if a packaged app was open at the time, the
downloads continued. With this change, on those platforms, when the last
browser window is closed, all in-progress downloads are cancelled.

BUG=320592

Review URL: https://codereview.chromium.org/73583005