Support use_glib==0 on Linux in base
This makes base build on Linux when use_glib is set to 0. The only glib
bits we are currently using in that configuration are g_get_home_dir()
and g_try_malloc().

Use of g_get_home_dir() is not essential. Without glib, we can just rely
on the $HOME environment variable always.

Use of g_try_malloc() is just a way to cause a large memory allocation
from a shared library. We can achieve the same effect using asprintf().

BUG=318315, 318413, 170408

Review URL: https://codereview.chromium.org/69953025