Use TOP #1 language from CLD2 instead of the summary language
The summary language and the top language are different, and the top language should be used for Translate. Originally, Chromium adopts the top language of CLD1, and summary language of CLD2. This CL fixes to use the top language even in the case of CLD2.

BUG=320362
TEST=manual

Review URL: https://codereview.chromium.org/75483009