Introduce RemoteInputMethodWin
In order to enable IME support on Ash, we need a proxy
object which implements ui::InputMethod interface and works
as if it was the actual IME back-end running in the
metro_driver process.

This CL introduces RemoteInputMethodWin for this purpose.
RemoteInputMethodWin is not used in the production code yet.
So this CL never breaks existing functionary at the moment.

BUG=164964
TEST=ui_unittests --gtest_filter=RemoteInputMethodTest.*

Review URL: https://codereview.chromium.org/67503004