Remove blacklisting of Mac OS 10.6 and 10.7.
Add blacklisting of VMWare driver on 10.6 and 10.7 because it hangs.

Disable compositing expectation test on 10.6 and 10.7 because it will
give different results depending on whether it is running on
a VM or real hardware.

Expand disabling of PhishingClassifierTest because it
flakes on all Mac versions now.

TBR=vangelis
BUG=266186

Review URL: https://codereview.chromium.org/66153005