Add JavaScript support to Mojo's bindings generator
This CL contains a first iteration of a JavaScript backend for Mojo's IPC
bindings generator. The generated code closely matches the hand-written
bindings for simple_service.js, although the generated message is off in a few
places.

This backend successfully generates code for hello_world_service.mojom and
math_calculator.mojom, but do not appear to be computing implicit ordinals yet,
which means these generated files not yet work. I'll fix these issues in the
next iteration.

Unlike the CPP backend, this JS backend uses jinja2 templates. If we're happy
with this approach, we'll likely port the CPP backend to jinja2 as well.

R=davemoore@chromium.org
BUG=317398

Review URL: https://codereview.chromium.org/77153003