Makes FocusdController honor focus change if no previously active window
With desktop aura when a Window is made active the focus restore code
kicks in and attempts to restore focus to the Window that had focus
prior to losing activation. FocusController ignores requests to change
focus if during the activation phase focus changed. Together this
means the window that initiated the focus and activation change does
not get focus.

Fix is to allow the focus change through if there was previously no
active window.

BUG=321042
TEST=covered by test now
R=ben@chromium.org

Review URL: https://codereview.chromium.org/75933005