telemetry: Rename new rasterize and record, add benchmark
This patch renames new_rasterize_and_record measurement to be
rasterize_and_record_micro, as well as adds a benchmark for it for
top_25 and key mobile sites.

It also adds an if block to skip the measurement on branches that
do not have the required chromium code.

NOTRY=True
R=ernstm@chromium.org, tonyg@chromium.org

Review URL: https://codereview.chromium.org/79093002