Track Dynamically Assigned Command IDs in Background Menu Generation
Menu items require that Command IDs be unique. For a single profile,
collisions don't happen. For multiple profiles, the profile menu
as well as its submenus start at Command ID 0, confusing the menu
system when it tries to identify or change and item.

The fix here creates an vector to hold the extension indexes
using the vector index as a Command ID. For profile menu labels,
the extension index is -1.

BUG=315164

Review URL: https://codereview.chromium.org/75973004