Add option to image_diff to calculate histogram diff
Currently, image_diff will return very high values for the percentage
diff in cases like the entire page getting shifted down by 1px.

This patch adds an alternative inexact way of computing the percentage
diff, which instead diffs the histograms of RGBA values to see whether
the same number of each color pixel occurs. Since the histogram ignores
at what position each pixel occurred, this diff is resistant to shifts,
and even rotation (if you just swap parts of the image, such that the
overall color distribution is the same, the histogram diff will be 0%).

An intended use for this is being able to sort a set of pixel diffs in
order of relative importance.

BUG=135823

Review URL: https://codereview.chromium.org/67973005