Fix focus issues with the Google earth NPAPI plugin in Windows desktop Aura.
The Google earth NPAPI plugin is a windowed plugin which sets focus to itself when we click on it. When we click on
the webpage after that the page never gets focus and all keystrokes continue to go to the plugin.

This works when we click on the omnibox because in its mouse down handler the omnibox eventually calls view::RequestFocus
which in turn calls the windows API SetFocus which sets keyboard focus to the window.

Fix is to do the same thing in the RenderwidgetHostViewaura::OnMouseEvent function.
The other change is in the HWNDMessageHandler::OnMouseActivate handler. We check if a child window is being activated and
if yes then we treat it as if the containing window lost activation. On a web page this would translate to Blur being sent
to the renderer etc. This is only done for AURA to eliminate this playing bad with non AURA code.

Fixes bug https://code.google.com/p/chromium/issues/detail?can=2&q=318488

BUG=318488
R=cpu@chromium.org, sky@chromium.org, jam, scottmg, sky

Review URL: https://codereview.chromium.org/78963004