Remove text media types from supported_non_image_types in mime_util.cc
The only user of the list is non_image_map_ and the only user of
non_image_map_ is IsSupportedNonImageMimeType(). This method has code
handling types starting with text/, so types not in
unsupported_text_types doesn't need to be included in
supported_non_image_types.

BUG=none

Review URL: https://codereview.chromium.org/78963005