Only supports EME when MSE is used in WebMediaPlayerAndroid.
Ideally, we should enable SRC ClearKey tests in content_browsertests and expect to get NOT_SUPPORTED_ERROR on Android. However, when SRC is used, the media pipeline will not wait for key to be added to start decoding. Therefore, we may get either NOT_SUPPORTED_ERROR or DECODE_ERROR. We can refactor media content browser tests to accept multiple expectations, or fix the tests to have a more deterministic behavior. But that should not be in the scope of this CL.

BUG=318886
TEST=none

Review URL: https://codereview.chromium.org/72113002