Omnibox: HistoryQuick Should Only Index & Search First 1k of URL & Title
Makes HistoryQuick provider only index the first 1k charcters in a URL
and the first 1k characters in a title.  This should save speed and
memory because the index will be smaller and faster to search when
users have long URLs in them.

I don't imagine there will be an quality impact because who searches
for a URL using characters so deep in the URL or the title?  I'm
so comfortable with this that I want to launch it without running an
omnibox experiment.

I bumped the HistoryQuick provider cache version number to enforce
that everyone rebuilds their cache.  This will let me easily look at
statistics about cache size and HQP speed before and after this change.
(Without bumping the cache version number, users will rebuild their
cache weekly, but I will not be able to easily tell which users have
rebuilt and which users have not, and hence cannot do a simple before/after
comparison.)

Tested to verify that HQP still seems to work.  Also tested that it
solves the lagginess described in the listed bug.

TBR=sky

BUG=281461

Review URL: https://codereview.chromium.org/81513002