Add ResourceType::PING and mark those requests as detachable.
This avoids cancelling <a ping> requests on cross-process navigations
when the owning renderer dies.

Actually assigning resources this type is conditioned on a currently
undefined WEBKIT_HAS_TARGET_PING symbol to be defined in a follow-up
Blink CL. The #ifdef can then be removed when that lands.

BUG=302816
TEST=ResourceDispatcherHostTest.DeletedFilterDetachable

Review URL: https://codereview.chromium.org/55513002