Android: Disable about:flags impl-side-painting override
We don't support running with impl-side painting disabled on Android
anymore. (It's hardcoded to enabled for Android in cc/base/switches.cc)

Removing it from the supported experiments should also cause it to be
ignored at startup when loading the profile (see ConvertFlagsToSwitches called
from ChromeBrowserMainParts::PreCreateThreadsImpl), so that users that turned
this off some time in the past will be forced to enabled again.

BUG=321545
TBR=thakis@chromium.org
NOTRY=True

Review URL: https://codereview.chromium.org/82553012