Point for WM_NCHITTEST should be in screen coordinates
In HWNDMessageHandler::OnMouseRange() function, mouse events synthesized
in response to touch should be ignored if they are in the client area of
the window. The hit testing to decide if a point is in the client area
needs the point in screen coordinates, but, for most of the mouse events
the point is in window coordinates. So, the point should be converted to
screen coordinates if it is not.

BUG=312708

Review URL: https://codereview.chromium.org/83973002